Refactored NavBar  and profile components tests to pass
