+++ src/components/pages/profile/EditProfileInputs.js
@@ -2,7 +2,7 @@ import React from "react";
 import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
 import Button from "@material-ui/core/Button";
-import SelectState from "../profile/SelectState";
+import SelectState from "./SelectState";
 import TextField from "@material-ui/core/TextField";
 import formStyles from "../profileStyles/formStyles";
